Version with live video trigger used in concert
In this version sax player 2 triggers video cues as well as metronome cues. A third person (it was me) was triggering the scenes in this setting. I do not own the clip I used in this concert, so I won't share it here, but it can be substituted with anything else.